chore: removing unused metrics
